seperate infobox and image parsing
