+++ src/parse/postProcess/index.js
@@ -1,0 +1,16 @@
+const i18n = require('../../data/i18n');
+const img_regex = new RegExp('^(' + i18n.images.concat(i18n.files).join('|') + ')', 'i');
+
+//cleanup after ourselves
+const postProcess = function(r) {
+  // add additional image from infobox, if applicable
+  if (r.infoboxes[0] && r.infoboxes[0].data && r.infoboxes[0].data['image'] && r.infoboxes[0].data['image'].text) {
+    let img = r.infoboxes[0].data['image'].text || '';
+    if (typeof img === 'string' && !img.match(img_regex)) {
+      img = 'File:' + img;
+    }
+    r.images.push(img);
+  }
+  return r;
+};
+module.exports = postProcess;
